Update employee competence card and pull out logic
