+++ backend/routers/employees/employeeChartConversion.ts
@@ -27,12 +27,14 @@ export const employeeMotivationAndCompetence = (
           indexBy,
           keys,
           data,
+          maxValue: 5,
         },
         {
           type: 'RadarChart',
           indexBy,
           keys,
           data,
+          maxValue: 5,
         },
       ],
     })),
